+++ src/utils/index.js
@@ -1,9 +1,11 @@
 "use strict";
 
+const cacheWrapper = require("./cache-wrapper");
 const getAfter = require("./get-after");
 const interpolateName = require("./interpolate-name");
 
 module.exports = {
+  cacheWrapper,
   getAfter,
   interpolateName
 };